requirements.txt
Django==2.2.2
PyMySQL==0.9.3
xadmin==2.0.1
django-formtools==2.1
django-crispy-forms==1.7.2
django-import-export==1.2.0
xlrd==1.2.0
XlsxWriter==1.1.8
xlwt==1.3.0
djangorestframework==3.10.2
requests==2.22.0
demjson==2.2.4
celery==4.3.0
django-celery-beat==1.5.0
django-celery-results ==1.1.2
redis==3.2.1
pyecharts==1.5.1